Add Biological Year to data

This function adds a column designating biological year starting on a user-designated date e.g. 1 June (day of year 152) for caribou

addBioYear(df, bioStart)

Arguments

df

a data frame with columns: Time: Date and time in POSIXct

bioStart

biological year start date, in day of year format yday

Value

Examples

data("caribou")

head(caribou)
#>       ID sex                Time Year       Lon      Lat         x       y
#> 1 Dancer   f 2002-04-01 00:00:00 2002 -133.8427 62.77028 -51589.97 7006600
#> 2 Dancer   f 2002-04-01 08:00:00 2002 -133.8394 62.76716 -51480.43 7006228
#> 3 Dancer   f 2002-04-01 16:00:00 2002 -133.8453 62.77767 -51584.51 7007437
#> 4 Dancer   f 2002-04-02 00:00:00 2002 -133.8643 62.80010 -52121.93 7010074
#> 5 Dancer   f 2002-04-02 08:00:00 2002 -133.8621 62.79918 -52024.50 7009953
#> 6 Dancer   f 2002-04-02 16:00:00 2002 -133.8625 62.80007 -52030.49 7010055

# bioStart is yday 152 (01 June)

caribou <- addBioYear(caribou, bioStart = 152)
head(caribou)
#>       ID sex                Time Year       Lon      Lat         x       y
#> 1 Dancer   f 2002-04-01 00:00:00 2002 -133.8427 62.77028 -51589.97 7006600
#> 2 Dancer   f 2002-04-01 08:00:00 2002 -133.8394 62.76716 -51480.43 7006228
#> 3 Dancer   f 2002-04-01 16:00:00 2002 -133.8453 62.77767 -51584.51 7007437
#> 4 Dancer   f 2002-04-02 00:00:00 2002 -133.8643 62.80010 -52121.93 7010074
#> 5 Dancer   f 2002-04-02 08:00:00 2002 -133.8621 62.79918 -52024.50 7009953
#> 6 Dancer   f 2002-04-02 16:00:00 2002 -133.8625 62.80007 -52030.49 7010055
#>   bioYear
#> 1    2001
#> 2    2001
#> 3    2001
#> 4    2001
#> 5    2001
#> 6    2001