get Daily Range

This function estimate (daily) Local Convex Hulls (LoCoH) by converting a simple feature into a LoCoH-xy object, identifying the nearest neighbors, creating a LoCoH-hullset object by constructing the local convex polygon using the nearest neighbors (nn) of each data point

getDailyRange(sf, crs, method = c("LoCoH", "KernelUD"), ...)

Arguments

sf

a simple feature containing (daily) GPS coordinates used to estimate the (daily) ranging area

crs

the crs of the coordinates in sf

method

either "LocoH" or "KernelUD", the method used to estimate the daily ranging area (see getLoCoH and getKernelUD for more details)

...

additional parameters to pass to kernelUD for the "KernelUD" method or lhs.iso.add for the "LoCoH" method

nn

(if method is "LoCoH") the number of nearest neighbors used to construct the local convex polygons

level

(if method is "LoCoH") A numeric value of the level of the isopleths (value comprised between .05 and 1, default is .95)

percent

(if method is "KernelUD") percent kernel (value comprised between 5 and 100, default is 95)

grid

(if method is "KernelUD") grid points (default 200, higher than adehabitat's)

Value

The function returns a simple feature of the isopleth(s), with the area

Examples

require(TuktuTools)

data(caribou)
CRS = "+proj=lcc +lat_1=50 +lat_2=70 +lat_0=65 +lon_0=-120 +x_0=0 +y_0=0 +ellps=GRS80 +datum=NAD83 +units=m +no_defs"

head(caribou)
#>       ID sex                Time Year       Lon      Lat         x       y
#> 1 Dancer   f 2002-04-01 00:00:00 2002 -133.8427 62.77028 -51589.97 7006600
#> 2 Dancer   f 2002-04-01 08:00:00 2002 -133.8394 62.76716 -51480.43 7006228
#> 3 Dancer   f 2002-04-01 16:00:00 2002 -133.8453 62.77767 -51584.51 7007437
#> 4 Dancer   f 2002-04-02 00:00:00 2002 -133.8643 62.80010 -52121.93 7010074
#> 5 Dancer   f 2002-04-02 08:00:00 2002 -133.8621 62.79918 -52024.50 7009953
#> 6 Dancer   f 2002-04-02 16:00:00 2002 -133.8625 62.80007 -52030.49 7010055

# for this example, we will take all the locations of the animals at a given day, 
# whatever the year, to have enough locations for this day.
# However, this method has been developped to estimate daily ranges for animals 
# monitored the same year.

# add the day of year
caribou <- caribou %>% mutate(yday = yday(Time))

# Take one day during winter and one day during summer to compare dispersion during those two seasons
days <- caribou %>% subset((mday(Time) == 15 & month(Time) == 7) |
                             (mday(Time) == 15 & month(Time) == 1))

# get average daily location for each individual and year then assign the same year for computing the daily range
days <- days %>% getDailyMean %>% st_as_sf(coords=c("x","y"), crs = CRS) %>% 
  mutate(Year = 2015, Date = as.character(as.Date(paste0("2015", substr(Time, 5, 10))))) %>% mutate(yday = yday(Date))

# daily LoCoH
require(tlocoh); require(raster)
#> Loading required package: tlocoh
#> Warning: there is no package called ‘tlocoh’
#> Loading required package: raster
#> 
#> Attaching package: ‘raster’
#> The following object is masked from ‘package:dplyr’:
#> 
#>     select
dailyranges.locoh <- getDailyRange(days, method = "LoCoH", nn = 5, level = .95, crs = CRS) %>%
  mutate(Date = as.Date(yday, origin = "2014-12-31") %>% as.character)
#> Error in mutate(., Date = as.Date(yday, origin = "2014-12-31") %>% as.character): ℹ In argument: `Date = as.Date(yday, origin = "2014-12-31") %>%
#>   as.character`.
#> Caused by error in `as.Date.default()`:
#> ! do not know how to convert 'yday' to class “Date”

# daily KernelUD
require(adehabitatHR)
#> Loading required package: adehabitatHR
#> Warning: there is no package called ‘adehabitatHR’
dailyranges.ud <- getDailyRange(days, method = "KernelUD", percent = 95) %>%
  mutate(Date = as.Date(yday, origin = "2014-12-31") %>% as.character)
#> Error in mutate(., Date = as.Date(yday, origin = "2014-12-31") %>% as.character): ℹ In argument: `Date = as.Date(yday, origin = "2014-12-31") %>%
#>   as.character`.
#> Caused by error in `as.Date.default()`:
#> ! do not know how to convert 'yday' to class “Date”

# visualize resulting daily ranges
map <- mapview::mapview(days, zcol = "Date")
m2 <- mapview::mapview(dailyranges.ud, map = map@map, zcol = "Date")
#> Error in eval(expr, envir, enclos): object 'dailyranges.ud' not found
mapview::mapview(dailyranges.locoh, map = m2@map, zcol = "Date")
#> Error in eval(expr, envir, enclos): object 'dailyranges.locoh' not found