get speed

This function calculate speed from raw movement data containing x and y coordinates (in meters) with corresponding date and time in POSIXct format. This will calculate the speed between subsequent relocations (i.e. of each step), as well as the middle time of the step.

getSpeed(x)

Arguments

x

a data frame containing columns: ID as individual identifiant, x and y coordinates (in a metric system) Time: date and time vector (of class POSIXct)

Value

The function returns the same dataframe than 'x' with columns `speed`, which is the movement rate between subsequent relocation (in m.h-1), `sl` which is the step length between subsequent locations, `dhours` which is an index of the cumulative time (in hour) from the first to the nth relocation, `dt` the time lag between subsequent relocation (in hour).

Examples

require(TuktuTools)

data(caribou)

# keep data during calving period (i.e., May 19 to July 7), no more than 3 consecutive days of missing data
# and at least 1 fix per day
b <- cutperiod(caribou, start = c(05,19), end = c(07,07), nfixes = 1, dayloss = 3)
#> Error in cutperiod(caribou, start = c(5, 19), end = c(7, 7), nfixes = 1,     dayloss = 3): could not find function "cutperiod"

# get movement rate
c <- getSpeed(b) %>% mutate(Year = year(Time))
#> Error in eval(expr, envir, enclos): object 'b' not found

# visualize movement rate

ggplot2::ggplot(data=c, aes(x=time.mid, y = speed, group = ID, color = ID)) + geom_path() + 
  facet_wrap(~Year, scales = "free")
#> Error in ggplot2::ggplot(data = c, aes(x = time.mid, y = speed, group = ID,     color = ID)): `data` cannot be a function.
#> ℹ Have you misspelled the `data` argument in `ggplot()`