Likelihood of Female having a calf

nllCalf calculates the negative log-likelihood of a female with a calf that survived the early rearing period

nllCalfdeath calculates the negative log-likelihood of a female with a calf that died.

nllCalf(df, BP1, kcons, PlotMe = FALSE)

nllCalfDeath(df, BP1, BP2, kcons, PlotMe = FALSE)

Arguments

df

dataframe from get.speed function

BP1

Break point corresponding to calf birth (in days from the begining of the dataset)

kcons

vector of the minimum and maximum time it takes the female to recover normal speed (in days)

PlotMe

if PlotMe = TRUE, a plot of the speed in function to the date and time will be drawn for each individual with the line of the speed estimated by the function with depending of BP1, BP2 and k

BP2

Break point corresponding to calf death (in days from the begining of the dataset)

Value

Returns the negative log-likelihood of this model and the recovery time (in days)

Details

The models assume that the mean speed is constant prior to the birth of a calf and after recovery or calf's death

Once the calf is born the speed decrease close to 0. Then the speed increases slowly until recovery or calf's death, then increases immediately to reach the initial speed before parturition if the calf died.

Once the mean speed reaches the value from before the birth of the calf, it remains at this value.

Functions

nllCalf(): Likelihood of Female having a calf that survived the

nllCalfDeath(): Likelihood of Female having a calf that died

References

DeMars, C., M. Auger-Méthé, U. Schlägel, S. Boutin, (2013) Inferring Parturition and Neonate Survival from Movement Patterns of Female Ungulates. Ecology and Evolution. DOI: 10.1002/ece3.785